pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-data-converter"
version = "0.1.0"
description = "Convert any document, text, or URL into LLM-ready data format"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nanonets", email = "team@nanonets.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "PyMuPDF>=1.23.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "openpyxl>=3.0.10",
    "pandas>=1.5.0",
    "Pillow>=9.0.0",
    "markdown>=3.4.0",
    "litellm>=1.0.0",
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
llm-converter = "llm_converter.cli:main"

[project.urls]
Homepage = "https://github.com/nanonets/llm-data-converter"
Repository = "https://github.com/nanonets/llm-data-converter"
Documentation = "https://github.com/nanonets/llm-data-converter#readme"
Issues = "https://github.com/nanonets/llm-data-converter/issues"

[tool.hatch.build.targets.wheel]
packages = ["llm_converter"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 